remove ambiguity for `max`, `min` w/ 2 seq's as args
In the current version of nim, there is a `min` and `max` proc in the
math module, which takes a generic parameter T (however, not excluding
seq or array). This means the previous definition using openArray as
the generic type in this module, would call the system.min
instead. That would fail. So as a workaround we use an explicit seq
for now.